Refactoring and adding label view
